Move some files and eliminate notebookCreationTracker (only used in serverpreload)
